19/12/19 gui refactor finish

+++ ml/_theano/dataloader.py
@@ -15,8 +15,8 @@ from _theano.tokenizer import *
 
 def load_data(data_dir: str, label_dir: str,
               tokenizer: Tokenizer = None,
-              valid_portion: int = 0.0,
-              test_portion: int = 0.1,
+              valid_portion: float = 0.0,
+              test_portion: float = 0.1,
               update_dict=True) -> (list, list, list):
     if label_dir is None:
         label_dict = None
